#6: refactor test setup
