+++ package.json
@@ -17,7 +17,7 @@
     "format": "prettier --write .",
     "check-format": "prettier --check .",
     "check-exports": "attw --pack . --ignore-rules=cjs-resolves-to-esm",
-    "test": "vitest run",
+    "test": "vitest --config tsconfig.vitest.json run",
     "local-release": "changeset version && changeset publish",
     "prepublishOnly": "npm run ci"
   },
